chore : cleaning code and renaming file to flutter standard
