+++ lib/ui/features/sirine/provider/location_stream_provider.dart
@@ -34,10 +34,7 @@ void updateRoutePolyline(
   );
 
   // Set the polyline (maybe just one of them, or both separately)
-  print(pointData);
-  print("FETCHED POINT DATA");
   if (pointData.containsKey('polyline')) {
-    print("GETTING POLYLINE DATA");
     final polyline = Polyline(
       polylineId: const PolylineId('route'),
       points: pointData['polyline'],
@@ -46,6 +43,5 @@ void updateRoutePolyline(
     );
 
     ref.read(routePolylineProvider.notifier).state = {polyline};
-    print("UPDATED POLYLINE");
   }
 }
