+++ lib/ui/features/sirine/provider/locked_destination_provider.dart
@@ -1,5 +1,4 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:founded_ninu/domain/entities/locked_address.dart';
-import 'package:google_maps_flutter/google_maps_flutter.dart';
 
 final lockedDestinationProvider = StateProvider<LockedAddress?>((ref) => null);
