+++ lib/ui/features/sirine/widgets/map_screen.dart
@@ -5,7 +5,6 @@ import 'package:founded_ninu/domain/use_cases/map_usecase.dart';
 import 'package:founded_ninu/ui/features/sirine/provider/cancel_confirmation_provider.dart';
 import 'package:founded_ninu/ui/features/sirine/provider/has_arrived_provider.dart';
 import 'package:founded_ninu/ui/features/sirine/provider/loading_provider.dart';
-import 'package:founded_ninu/ui/features/sirine/provider/location_provider.dart';
 import 'package:founded_ninu/ui/features/sirine/provider/location_stream_provider.dart';
 import 'package:founded_ninu/ui/features/sirine/provider/overlay_prompt_provider.dart';
 import 'package:founded_ninu/ui/features/sirine/widgets/cancel_confirmation.dart';
