+++ lib/ui/features/sirine/widgets/second_start_mode_bottom_sheet.dart
@@ -7,7 +7,6 @@ import 'package:founded_ninu/ui/features/sirine/provider/radiating_sirine_icon.d
 import 'package:founded_ninu/ui/features/sirine/provider/bottomsheet_provider.dart';
 import 'package:founded_ninu/ui/features/sirine/provider/scaffold_provider.dart';
 import 'package:founded_ninu/ui/features/sirine/widgets/first_start_mode_bottom_sheet.dart';
-import 'package:go_router/go_router.dart';
 
 class SecondBottomSheet extends ConsumerStatefulWidget {
   const SecondBottomSheet({super.key});
